+++ host/modules/api.jsx
@@ -1,5 +1,7 @@
 $.global.FishTools = {
     executeTool: executeTool,
     readEase: _readEase,
-    applyEase: _applyEase
+    applyEase: _applyEase,
+    readVelocity: _readVelocity,
+    applyVelocity: _applyVelocity
 };